+++ source/transmission_line/cpw.py
@@ -1,7 +1,9 @@
 """This module contains equations for calculating properties of co-planar waveguide transmission lines, such as their
 inductance and capacitance per unit length, and classes for drawing them as GDSII structures.
 
-Most of the classes draw structures with **negative polarity**, meaning that structures represent the absence of metal.
+The classes with name that start with 'Positive' draw the positive space, meaning that structures correspond to metal.
+The classes with names that start with 'Negative' draw the negative space, meaning that structures correspond to the
+absence of metal.
 """
 from __future__ import absolute_import, division, print_function
 
@@ -273,7 +275,7 @@ class AbstractCPW(AbstractTransmissionLine):
 
     @property
     def inductance_per_unit_length(self):
-        """Return the total (geometric + kinetic, if thickness is given) inductance per unit length."""
+        """Return the total (geometric + kinetic, if thickness was given) inductance per unit length."""
         if self.thickness is None:
             return self.geometric_inductance_per_unit_length
         else:
@@ -281,12 +283,13 @@ class AbstractCPW(AbstractTransmissionLine):
 
 
 # ToDo: determine how to handle the multiple inheritance for AbstractCPW and SmoothedSegment
+# Classes that draw the negative space of CPW structures.
 
-class CPW(SmoothedSegment):
-    """The negative space of a segment of co-planar waveguide."""
+class NegativeCPW(SmoothedSegment):
+    """The negative space of a segment of co-planar waveguide: structures are absence of metal."""
 
     def __init__(self, outline, trace, gap, radius=None, points_per_radian=DEFAULT_POINTS_PER_RADIAN, round_to=None):
-        """Instantiating a CPW does not draw it into any cell.
+        """Instantiate without drawing in any cell.
 
         :param outline: the vertices of the CPW path, before smoothing; see :func:`smooth`.
         :param float trace: the width of the center trace.
@@ -300,8 +303,8 @@ class CPW(SmoothedSegment):
         self.gap = gap
         if radius is None:
             radius = trace / 2 + gap
-        super(CPW, self).__init__(outline=outline, radius=radius, points_per_radian=points_per_radian,
-                                  round_to=round_to)
+        super(NegativeCPW, self).__init__(outline=outline, radius=radius, points_per_radian=points_per_radian,
+                                          round_to=round_to)
 
     def draw(self, cell, origin, layer, datatype=0):
         """Draw this structure into the given cell and return the drawn polygon set, which should contain two polygons.
@@ -321,14 +324,16 @@ class CPW(SmoothedSegment):
         return polygon_set
 
 
-class CPWBlank(SmoothedSegment):
-    """Negative co-planar waveguide with the center trace missing, that is, the negative space of the trace and gaps.
+class NegativeCPWBlank(SmoothedSegment):
+    """Negative co-planar waveguide with the center trace missing, that is, the negative space of the trace and gaps:
+    structures are absence of metal.
 
-     This is useful when the center trace is on a separate layer or has a separate datatype.
-     """
+    This class draws a single FlexPath that is the union of the trace and gaps. This is useful when the trace is on
+    another layer or has a different datatype. The interface is the same as :class:`CPW` for compatibility.
+    """
 
     def __init__(self, outline, trace, gap, radius=None, points_per_radian=DEFAULT_POINTS_PER_RADIAN, round_to=None):
-        """Instantiating a CPWBlank does not draw it into any cell.
+        """Instantiate without drawing in any cell.
 
         :param outline: the vertices of the CPW path, before smoothing; see :func:`smooth`.
         :param float trace: the width of the center trace.
@@ -342,8 +347,8 @@ class CPWBlank(SmoothedSegment):
         self.gap = gap
         if radius is None:
             radius = trace / 2 + gap
-        super(CPWBlank, self).__init__(outline=outline, radius=radius, points_per_radian=points_per_radian,
-                                       round_to=round_to)
+        super(NegativeCPWBlank, self).__init__(outline=outline, radius=radius, points_per_radian=points_per_radian,
+                                               round_to=round_to)
 
     def draw(self, cell, origin, layer, datatype=0):
         """Draw this structure into the given cell and return the drawn polygon set, which should contain one polygon.
@@ -362,13 +367,13 @@ class CPWBlank(SmoothedSegment):
         return path_set
 
 
-class CPWElbowCoupler(SmoothedSegment):
-    """Negative co-planar waveguide elbow coupler."""
+class NegativeCPWElbowCoupler(SmoothedSegment):
+    """Negative co-planar waveguide elbow coupler: structures are absence of metal."""
 
     def __init__(self, tip_point, elbow_point, joint_point, trace, gap, radius=None,
                  points_per_radian=DEFAULT_POINTS_PER_RADIAN,
                  round_to=None):
-        """Instantiating a CPWElbowCoupler does not draw it into any cell.
+        """Instantiate without drawing in any cell.
 
         :param point tip_point: the open end of the coupler; the first point of the segment.
         :param point elbow_point: the point where the coupler turns away from the feedline; the middle point of the
@@ -386,8 +391,8 @@ class CPWElbowCoupler(SmoothedSegment):
         self.gap = gap
         if radius is None:
             radius = trace / 2 + gap
-        super(CPWElbowCoupler, self).__init__(outline=[tip_point, elbow_point, joint_point], radius=radius,
-                                              points_per_radian=points_per_radian, round_to=round_to)
+        super(NegativeCPWElbowCoupler, self).__init__(outline=[tip_point, elbow_point, joint_point], radius=radius,
+                                                      points_per_radian=points_per_radian, round_to=round_to)
 
     def draw(self, cell, origin, layer, datatype=0, round_tip=True):
         """Draw this structure into the given cell and return the drawn polygon set, which should contain two polygons.
@@ -419,16 +424,18 @@ class CPWElbowCoupler(SmoothedSegment):
         return path_set, round_set
 
 
-class CPWElbowCouplerBlank(SmoothedSegment):
+class NegativeCPWElbowCouplerBlank(SmoothedSegment):
     """Negative co-planar waveguide elbow coupler with the center trace missing, that is, the negative space of the
-    trace and gaps.
+    trace and gaps: structures are absence of metal.
 
-     This is useful when the center trace is on a separate layer or has a separate datatype.
-     """
+    This class draws a single polygon that is the union of the trace and gaps. This is useful when the trace is on
+    another layer or has a different datatype. The interface is the same as :class:`NegativeCPWElbowCoupler` for
+    compatibility.
+    """
 
     def __init__(self, tip_point, elbow_point, joint_point, trace, gap, radius=None,
                  points_per_radian=DEFAULT_POINTS_PER_RADIAN, round_to=None):
-        """Instantiating a CPWElbowCouplerBlank does not draw it into any cell.
+        """Instantiate without drawing in any cell.
 
         :param point tip_point: the open end of the coupler; the first point of the segment.
         :param point elbow_point: the point where the coupler turns away from the feedline; the middle point of the
@@ -446,8 +453,8 @@ class CPWElbowCouplerBlank(SmoothedSegment):
         self.gap = gap
         if radius is None:
             radius = trace / 2 + gap
-        super(CPWElbowCouplerBlank, self).__init__(outline=[tip_point, elbow_point, joint_point], radius=radius,
-                                                   points_per_radian=points_per_radian, round_to=round_to)
+        super(NegativeCPWElbowCouplerBlank, self).__init__(outline=[tip_point, elbow_point, joint_point], radius=radius,
+                                                           points_per_radian=points_per_radian, round_to=round_to)
 
     def draw(self, cell, origin, layer, datatype=0, round_tip=True):
         """Draw this structure into the given cell and return the drawn polygon set, which should contain two polygons.
@@ -476,11 +483,14 @@ class CPWElbowCouplerBlank(SmoothedSegment):
         return path_set, round_set
 
 
-class CPWTransition(Segment):
-    """Negative transition between two sections of co-planar waveguide."""
+class NegativeCPWTransition(Segment):
+    """Negative transition between two sections of co-planar waveguide: structures are absence of metal.
+
+    The points of this structure are (start_point, end_point).
+    """
 
     def __init__(self, start_point, start_trace, start_gap, end_trace, end_point, end_gap, round_to=None):
-        """Instantiating a CPWTransition does not draw it into any cell.
+        """Instantiate without drawing in any cell.
 
         The points of this structure are (start_point, end_point).
 
@@ -492,7 +502,7 @@ class CPWTransition(Segment):
         :param end_gap: the gap width of the following section.
         :param round_to: see :class:`SmoothedSegment`.
         """
-        super(CPWTransition, self).__init__(points=[start_point, end_point], round_to=round_to)
+        super(NegativeCPWTransition, self).__init__(points=[start_point, end_point], round_to=round_to)
         self.start_trace = start_trace
         self.start_gap = start_gap
         self.end_trace = end_trace
@@ -528,12 +538,18 @@ class CPWTransition(Segment):
         return upper_polygon, lower_polygon
 
 
-class CPWTransitionBlank(Segment):
-    """Negative transition between two sections of co-planar waveguide, used when the center trace is a separate
-    layer."""
+class NegativeCPWTransitionBlank(Segment):
+    """Negative transition between two sections of co-planar waveguide: structures are absence of metal.
+
+    The points of this structure are (start_point, end_point).
+
+    This class draws a single polygon that is the union of the trace and gaps. This is useful when the trace is on
+    another layer or has a different datatype. The interface is the same as :class:`NegativeCPWTransition` for
+    compatibility.
+    """
 
     def __init__(self, start_point, start_trace, start_gap, end_point, end_trace, end_gap, round_to=None):
-        """Instantiating a CPWTransitionBlank does not draw it into any cell.
+        """Instantiate without drawing in any cell.
 
         The points of this structure are (start_point, end_point).
 
@@ -545,7 +561,7 @@ class CPWTransitionBlank(Segment):
         :param end_gap: the gap width of the following section.
         :param round_to: see :class:`SmoothedSegment`.
         """
-        super(CPWTransitionBlank, self).__init__(points=[start_point, end_point], round_to=round_to)
+        super(NegativeCPWTransitionBlank, self).__init__(points=[start_point, end_point], round_to=round_to)
         self.start_trace = start_trace
         self.start_gap = start_gap
         self.end_trace = end_trace
@@ -575,282 +591,167 @@ class CPWTransitionBlank(Segment):
         cell.add(element=polygon)
         return polygon
 
+# Classes that draw the positive space of CPW structures with finite ground planes.
 
-# ToDo: split this into separate classes
-# ToDo: refactor this so that it's clear that how to mix it with Segment subclasses
-class Mesh(object):
-    """A mix-in class that allows Segment subclasses that have the same outlines to share mesh code."""
+class PositiveCPW(SmoothedSegment):
+    """The positive space of a segment of co-planar waveguide with finite ground planes: structures are metal."""
 
-    def path_mesh(self):
-        """Return the centers of the mesh elements calculated using the Segment's parameters.
+    def __init__(self, outline, trace, gap, ground, radius=None, points_per_radian=DEFAULT_POINTS_PER_RADIAN,
+                 round_to=None):
+        """Instantiate without drawing in any cell.
 
-        :return: the centers of the mesh elements.
-        :rtype: list[numpy.ndarray]
+        :param list[point] outline: the vertices of the CPW path, before smoothing; see :func:`smooth`.
+        :param float trace: the width of the center trace.
+        :param float gap: the width of the gaps on each side of the center trace between it and the ground planes.
+        :param float ground: the width of the ground planes.
+        :param radius: see :func:`smooth`.
+        :type radius: float or None
+        :param int points_per_radian: see :func:`smooth`.
+        :param float round_to: see :class:`SmoothedSegment`.
         """
-        mesh_centers = []
-        center_to_first_row = self.trace / 2 + self.gap + self.mesh_border
-        # Mesh the straight sections
-        starts = [self.start] + [bend[-1] for bend in self.bends]
-        ends = [bend[0] for bend in self.bends] + [self.end]
-        for start, end in zip(starts, ends):
-            v = end - start
-            length = np.linalg.norm(v)
-            phi = np.arctan2(v[1], v[0])
-            R = np.array([[np.cos(phi), -np.sin(phi)],
-                          [np.sin(phi), np.cos(phi)]])
-            num_mesh_columns = int(np.floor(length / self.mesh_spacing))
-            if num_mesh_columns == 0:
-                continue
-            elif num_mesh_columns == 1:
-                x = np.array([length / 2])
-            else:
-                x = np.linspace(self.mesh_spacing / 2, length - self.mesh_spacing / 2, num_mesh_columns)
-            y = center_to_first_row + self.mesh_spacing * np.arange(self.num_mesh_rows)
-            xx, yy = np.meshgrid(np.concatenate((x, x)), np.concatenate((y, -y)))
-            Rxy = np.dot(R, np.vstack((xx.flatten(), yy.flatten())))
-            mesh_centers.extend(zip(start[0] + Rxy[0, :], start[1] + Rxy[1, :]))
-        # Mesh the curved sections
-        for row in range(self.num_mesh_rows):
-            center_to_row = center_to_first_row + row * self.mesh_spacing
-            for radius in [self.radius - center_to_row, self.radius + center_to_row]:
-                if radius < self.mesh_spacing / 2:
-                    continue
-                for angle, corner, offset in zip(self.angles, self.corners, self.offsets):
-                    num_points = int(np.round(radius * np.abs(angle) / self.mesh_spacing))
-                    if num_points == 1:
-                        max_angle = 0
-                    else:
-                        max_angle = (1 - 1 / num_points) * angle / 2
-                    mesh_centers.extend([corner + offset +
-                                         radius * np.array([np.cos(phi), np.sin(phi)]) for phi in
-                                         (np.arctan2(-offset[1], -offset[0]) +
-                                          np.linspace(-max_angle, max_angle, num_points))])
-        return mesh_centers
+        self.trace = trace
+        self.gap = gap
+        self.ground = ground
+        if radius is None:
+            radius = trace / 2 + gap
+        super(PositiveCPW, self).__init__(outline=outline, radius=radius, points_per_radian=points_per_radian,
+                                          round_to=round_to)
 
-    def trapezoid_mesh(self):
-        """
+    def draw(self, cell, origin, layer, datatype=0):
+        """Draw this structure into the given cell and return the drawn polygon set, which contains three polygons.
 
-        :return:
+        :param gdspy.Cell cell: the cell into which to draw the structure.
+        :param point origin: the points of the drawn structure are relative to this point.
+        :param int layer: the layer on which to draw.
+        :param int datatype: the GDSII datatype.
+        :return: the object that was drawn into the cell.
+        :rtype: gdspy.PolygonSet
         """
-        mesh_centers = []
+        points = [to_point(origin) + point for point in self.points]
+        trace_flexpath = gdspy.FlexPath(points=points, width=self.trace, max_points=0, gdsii_path=False)
+        gap_flexpath = gdspy.FlexPath(points=points, width=self.trace + 2 * self.gap, max_points=0,
+                                      gdsii_path=False)
+        ground_flexpath = gdspy.FlexPath(points=points, width=self.trace + 2 * self.gap + 2 * self.ground, max_points=0,
+                                         gdsii_path=False)
+        negative_polygon_set = gdspy.boolean(gap_flexpath, trace_flexpath, 'not', max_points=0, layer=layer,
+                                             datatype=datatype)
+        positive_polygon_set = gdspy.boolean(ground_flexpath, negative_polygon_set, 'not', max_points=0, layer=layer,
+                                             datatype=datatype)
+        cell.add(element=positive_polygon_set)
+        return positive_polygon_set
+
+
+class PositiveCPWBlank(SmoothedSegment):
+    """The positive space of a segment of co-planar waveguide with finite ground planes, with the trace removed:
+    structures are metal.
+
+    This class draws only the ground planes, with no center trace. This is useful when the trace is on another layer or
+    has a different datatype. The interface is the same as :class:`PositiveCPW` for compatibility.
+    """
+
+    def __init__(self, outline, trace, gap, ground, radius=None, points_per_radian=DEFAULT_POINTS_PER_RADIAN,
+                 round_to=None):
+        """Instantiate without drawing in any cell.
+
+        :param list[point] outline: the vertices of the CPW path, before smoothing; see :func:`smooth`.
+        :param float trace: the width of the center trace.
+        :param float gap: the width of the gaps on each side of the center trace between it and the ground planes.
+        :param float ground: the width of the ground planes.
+        :param radius: see :func:`smooth`.
+        :type radius: float or None
+        :param int points_per_radian: see :func:`smooth`.
+        :param float round_to: see :class:`SmoothedSegment`.
+        """
+        self.trace = trace
+        self.gap = gap
+        self.ground = ground
+        if radius is None:
+            radius = trace / 2 + gap
+        super(PositiveCPWBlank, self).__init__(outline=outline, radius=radius, points_per_radian=points_per_radian,
+                                               round_to=round_to)
+
+    def draw(self, cell, origin, layer, datatype=0):
+        """Draw this structure into the given cell and return the drawn polygon set, which contains two polygons.
+
+        :param gdspy.Cell cell: the cell into which to draw the structure.
+        :param point origin: the points of the drawn structure are relative to this point.
+        :param int layer: the layer on which to draw.
+        :param int datatype: the GDSII datatype.
+        :return: the object that was drawn into the cell.
+        :rtype: gdspy.PolygonSet
+        """
+        points = [to_point(origin) + point for point in self.points]
+        trace_and_gap_flexpath = gdspy.FlexPath(points=points, width=self.trace + 2 * self.gap, max_points=0,
+                                                gdsii_path=False)
+        ground_flexpath = gdspy.FlexPath(points=points, width=self.trace + 2 * self.gap + 2 * self.ground, max_points=0,
+                                         gdsii_path=False)
+        polygon_set = gdspy.boolean(ground_flexpath, trace_and_gap_flexpath, 'not', max_points=0, layer=layer,
+                                    datatype=datatype)
+        cell.add(element=polygon_set)
+        return polygon_set
+
+
+class PositiveCPWTransition(Segment):
+    """Positive transition between two sections of co-planar waveguide."""
+
+    def __init__(self, start_point, start_trace, start_gap, start_ground, end_trace, end_point, end_gap, end_ground,
+                 round_to=None):
+        """Instantiate without drawing in any cell.
+
+        The points of this structure are (start_point, end_point).
+
+        :param point start_point: the start point of the transition, typically the end point of the previous section.
+        :param start_trace: the trace width of the previous section.
+        :param start_gap: the gap width of the previous section.
+        :param start_ground: the width of both of the ground planes of the previous section.
+        :param end_point: the end point of the transition, typically the start point of the following section.
+        :param end_trace: the trace width of the following section.
+        :param end_gap: the gap width of the following section.
+        :param end_ground: the width of both of the ground planes of the following section.
+        :param round_to: see :class:`SmoothedSegment`.
+        """
+        super(PositiveCPWTransition, self).__init__(points=[start_point, end_point], round_to=round_to)
+        self.start_trace = start_trace
+        self.start_gap = start_gap
+        self.start_ground = start_ground
+        self.end_trace = end_trace
+        self.end_gap = end_gap
+        self.end_ground = end_ground
+
+    def draw(self, cell, origin, layer, datatype=0):
+        """Draw this structure into the given cell and return the three drawn polygons, which are the center trace and
+        two ground planes.
+
+        :param gdspy.Cell cell: the cell into which to draw the structure.
+        :param point origin: the points of the drawn structure are relative to this point.
+        :param int layer: the layer on which to draw.
+        :param int datatype: the GDSII datatype.
+        :return: the three polygons that were drawn into the cell.
+        :rtype: tuple[gdspy.Polygon]
+        """
         v = self.end - self.start
-        length = np.linalg.norm(v)
         phi = np.arctan2(v[1], v[0])
-        R = np.array([[np.cos(phi), -np.sin(phi)],
-                      [np.sin(phi), np.cos(phi)]])
-        start_to_first_row = self.start_trace / 2 + self.start_gap + self.start_mesh_border
-        difference_to_first_row = self.end_trace / 2 + self.end_gap + self.end_mesh_border - start_to_first_row
-        num_mesh_columns = int(np.floor(length / self.mesh_spacing))
-        if num_mesh_columns == 0:
-            return []
-        elif num_mesh_columns == 1:
-            x = np.array([length / 2])
-        else:
-            x = np.linspace(self.mesh_spacing / 2, length - self.mesh_spacing / 2, num_mesh_columns)
-        y = self.mesh_spacing * np.arange(self.num_mesh_rows, dtype=np.float)
-        xxp, yyp = np.meshgrid(x, y)  # These correspond to the positive y-values
-        y_shift = start_to_first_row + difference_to_first_row * x / length
-        yyp += y_shift
-        xx = np.concatenate((xxp, xxp))
-        yy = np.concatenate((yyp, -yyp))  # The negative y-values are reflected
-        Rxy = np.dot(R, np.vstack((xx.flatten(), yy.flatten())))
-        mesh_centers.extend(zip(self.start[0] + Rxy[0, :], self.start[1] + Rxy[1, :]))
-        return mesh_centers
-
-
-class CPWMesh(CPW, Mesh):
-    """doc me!"""
-
-    def __init__(self, outline, trace, gap, mesh_spacing, mesh_border, mesh_radius, num_circle_points, num_mesh_rows,
-                 radius=None, points_per_radian=DEFAULT_POINTS_PER_RADIAN, round_to=None):
-        """
-
-        :param outline:
-        :param width:
-        :param gap:
-        :param mesh_spacing:
-        :param mesh_border:
-        :param mesh_radius:
-        :param num_circle_points:
-        :param num_mesh_rows:
-        :param radius:
-        :param points_per_radian:
-        :param round_to:
-        """
-        super(CPWMesh, self).__init__(outline=outline, width=width, gap=gap, radius=radius,
-                                      points_per_radian=points_per_radian, round_to=round_to)
-        self.mesh_spacing = mesh_spacing
-        self.mesh_radius = mesh_radius
-        self.mesh_border = mesh_border
-        self.num_circle_points = num_circle_points
-        self.num_mesh_rows = num_mesh_rows
-        self.mesh_centers = self.path_mesh()
-
-    def draw(self, cell, origin, positive_layer, negative_layer, result_layer):
-        """
-
-        :param cell:
-        :param origin:
-        :param positive_layer:
-        :param negative_layer:
-        :param result_layer:
-        :return:
-        """
-        super(CPWMesh, self).draw(cell=cell, origin=origin, positive_layer=positive_layer,
-                                  negative_layer=negative_layer, result_layer=result_layer)
-        for mesh_center in self.mesh_centers:
-            cell.add_circle(origin=origin + mesh_center, radius=self.mesh_radius, layer=result_layer,
-                            number_of_points=self.num_circle_points)
-
-
-class CPWBlankMesh(CPWBlank, Mesh):
-    """doc me!"""
-
-    def __init__(self, outline, width, gap, mesh_spacing, mesh_border, mesh_radius, num_circle_points, num_mesh_rows,
-                 radius=None, points_per_radian=DEFAULT_POINTS_PER_RADIAN, round_to=None):
-        """
-
-        :param outline:
-        :param width:
-        :param gap:
-        :param mesh_spacing:
-        :param mesh_border:
-        :param mesh_radius:
-        :param num_circle_points:
-        :param num_mesh_rows:
-        :param radius:
-        :param points_per_radian:
-        :param round_to:
-        """
-        super(CPWBlankMesh, self).__init__(outline=outline, width=width, gap=gap, radius=radius,
-                                           points_per_radian=points_per_radian, round_to=round_to)
-        self.mesh_spacing = mesh_spacing
-        self.mesh_radius = mesh_radius
-        self.mesh_border = mesh_border
-        self.num_circle_points = num_circle_points
-        self.num_mesh_rows = num_mesh_rows
-        self.mesh_centers = self.path_mesh()
-
-    def draw(self, cell, origin, positive_layer, negative_layer, result_layer):
-        """
-
-        :param cell:
-        :param origin:
-        :param positive_layer:
-        :param negative_layer:
-        :param result_layer:
-        :return:
-        """
-        super(CPWBlankMesh, self).draw(cell=cell, origin=origin, positive_layer=positive_layer,
-                                       negative_layer=negative_layer, result_layer=result_layer)
-        for mesh_center in self.mesh_centers:
-            cell.add_circle(origin=origin + mesh_center, radius=self.mesh_radius, layer=result_layer,
-                            number_of_points=self.num_circle_points)
-
-
-class CPWElbowCouplerMesh(CPWElbowCoupler):
-    """doc me!"""
-    pass
-
-
-class CPWElbowCouplerBlankMesh(CPWElbowCouplerBlank):
-    """doc me!"""
-    pass
-
-
-class CPWTransitionMesh(CPWTransition, Mesh):
-    """doc me!"""
-
-    def __init__(self, start_point, end_point, start_width, end_width, start_gap, end_gap, mesh_spacing,
-                 start_mesh_border, end_mesh_border, mesh_radius, num_circle_points, num_mesh_rows, round_to=None):
-        """
-
-        :param start_point:
-        :param end_point:
-        :param start_width:
-        :param end_width:
-        :param start_gap:
-        :param end_gap:
-        :param mesh_spacing:
-        :param start_mesh_border:
-        :param end_mesh_border:
-        :param mesh_radius:
-        :param num_circle_points:
-        :param num_mesh_rows:
-        :param round_to:
-        """
-        super(CPWTransitionMesh, self).__init__(start_point=start_point, end_point=end_point, start_width=start_width,
-                                                end_width=end_width, start_gap=start_gap, end_gap=end_gap,
-                                                round_to=round_to)
-        self.mesh_spacing = mesh_spacing
-        self.start_mesh_border = start_mesh_border
-        self.end_mesh_border = end_mesh_border
-        self.mesh_radius = mesh_radius
-        self.num_circle_points = num_circle_points
-        self.num_mesh_rows = num_mesh_rows
-        self.mesh_centers = self.trapezoid_mesh()
-
-    def draw(self, cell, origin, positive_layer, negative_layer, result_layer):
-        """
-
-        :param cell:
-        :param origin:
-        :param positive_layer:
-        :param negative_layer:
-        :param result_layer:
-        :return:
-        """
-        super(CPWTransitionMesh, self).draw(cell=cell, origin=origin, positive_layer=positive_layer,
-                                            negative_layer=negative_layer, result_layer=result_layer)
-        for mesh_center in self.mesh_centers:
-            cell.add_circle(origin=origin + mesh_center, radius=self.mesh_radius, layer=result_layer,
-                            number_of_points=self.num_circle_points)
-
-
-class CPWTransitionBlankMesh(CPWTransitionBlank, Mesh):
-    """doc me!"""
-
-    def __init__(self, start_point, end_point, start_width, end_width, start_gap, end_gap, mesh_spacing,
-                 start_mesh_border, end_mesh_border, mesh_radius, num_circle_points, num_mesh_rows, round_to=None):
-        """
-
-        :param start_point:
-        :param end_point:
-        :param start_width:
-        :param end_width:
-        :param start_gap:
-        :param end_gap:
-        :param mesh_spacing:
-        :param start_mesh_border:
-        :param end_mesh_border:
-        :param mesh_radius:
-        :param num_circle_points:
-        :param num_mesh_rows:
-        :param round_to:
-        """
-        super(CPWTransitionBlankMesh, self).__init__(start_point=start_point, end_point=end_point,
-                                                     start_width=start_width, end_width=end_width,
-                                                     start_gap=start_gap, end_gap=end_gap, round_to=round_to)
-        self.mesh_spacing = mesh_spacing
-        self.start_mesh_border = start_mesh_border
-        self.end_mesh_border = end_mesh_border
-        self.mesh_radius = mesh_radius
-        self.num_circle_points = num_circle_points
-        self.num_mesh_rows = num_mesh_rows
-        self.mesh_centers = self.trapezoid_mesh()
-
-    def draw(self, cell, origin, positive_layer, negative_layer, result_layer):
-        """
-
-        :param cell:
-        :param origin:
-        :param positive_layer:
-        :param negative_layer:
-        :param result_layer:
-        :return:
-        """
-        super(CPWTransitionBlankMesh, self).draw(cell=cell, origin=origin, positive_layer=positive_layer,
-                                                 negative_layer=negative_layer, result_layer=result_layer)
-        for mesh_center in self.mesh_centers:
-            cell.add_circle(origin=origin + mesh_center, radius=self.mesh_radius, layer=result_layer,
-                            number_of_points=self.num_circle_points)
+        rotation = np.array([[np.cos(phi), -np.sin(phi)],
+                             [np.sin(phi), np.cos(phi)]])
+        trace = [(0, self.start_trace / 2),
+                 (self.length, self.end_trace / 2),
+                 (self.length, -self.end_trace / 2),
+                 (0, -self.start_trace / 2)]
+        upper = [(0, self.start_trace / 2 + self.start_gap),
+                 (0, self.start_trace / 2 + self.start_gap + self.start_ground),
+                 (self.length, self.end_trace / 2 + self.end_gap + self.end_ground),
+                 (self.length, self.end_trace / 2 + self.end_gap)]
+        lower = [(x, -y) for x, y in upper]
+        trace_rotated = [np.dot(rotation, to_point(p).T).T for p in trace]
+        upper_rotated = [np.dot(rotation, to_point(p).T).T for p in upper]
+        lower_rotated = [np.dot(rotation, to_point(p).T).T for p in lower]
+        trace_rotated_shifted = [to_point(origin) + self.start + p for p in trace_rotated]
+        upper_rotated_shifted = [to_point(origin) + self.start + p for p in upper_rotated]
+        lower_rotated_shifted = [to_point(origin) + self.start + p for p in lower_rotated]
+        trace_polygon = gdspy.Polygon(points=trace_rotated_shifted, layer=layer, datatype=datatype)
+        upper_polygon = gdspy.Polygon(points=upper_rotated_shifted, layer=layer, datatype=datatype)
+        lower_polygon = gdspy.Polygon(points=lower_rotated_shifted, layer=layer, datatype=datatype)
+        cell.add(element=trace_polygon)
+        cell.add(element=upper_polygon)
+        cell.add(element=lower_polygon)
+        return trace_polygon,upper_polygon, lower_polygon