+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021-03-05
+### Added
+- Three PositiveCPW classes for drawing the positive space of CPWs with finite ground planes.
+
+### Changed 
+- Renamed all previous classes starting with "CPW" to "NegativeCPW".
+
+### Removed
+- All Mesh classes in cpw.py; use mesh.py instead.
+
 ## [0.2.0] - 2021-03-04
 ### Changed
 - CPW center trace variable 'width' is now 'trace' throughout.